feat(proxy): Arb Commitment Type && invalid cert signal - fix header byte test

+++ api/proxy/test/e2e/server_arb_test.go
@@ -35,7 +35,7 @@ func TestArbCustomDAGetSupportedHeaderBytesMethod(t *testing.T) {
 	err = rpcClient.Call(&supportedHeaderBytesResult,
 		arbitrum_altda.MethodGetSupportedHeaderBytes)
 	require.NoError(t, err)
-	require.Equal(t, supportedHeaderBytesResult.HeaderBytes[0], commitments.ArbCustomDAHeaderByte)
+	require.Equal(t, supportedHeaderBytesResult.HeaderBytes[0], uint8(commitments.ArbCustomDAHeaderByte))
 
 }
